# src/docs/swagger.yaml
definitions:
  github_com_wahid-wex_e-commerce-backend_api_dto.AddRemoveToFavoriteRequest:
    properties:
      product_id:
        type: integer
    required:
    - product_id
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.CartItemResponse:
    properties:
      imageUrl:
        type: string
      name:
        type: string
      price:
        type: number
      productId:
        type: integer
      quantity:
        type: integer
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      name:
        type: string
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.CreateReviewRequest:
    properties:
      content:
        maxLength: 1000
        minLength: 1
        type: string
      product-id:
        type: integer
      rating:
        type: integer
    required:
    - content
    - product-id
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateProductRequest:
    properties:
      attributes:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductAttributeDTO'
        type: array
      category_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 200
        minLength: 1
        type: string
      price:
        type: number
      stocks:
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateProductStockRequest:
    properties:
      productId:
        type: integer
      quantity:
        type: integer
    required:
    - productId
    - quantity
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.FavoriteResponse:
    properties:
      customerId:
        type: integer
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        type: string
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.PagedList-github_com_wahid-wex_e-commerce-backend_api_dto_ProductResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.Sort'
        type: array
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.ProductAttributeDTO:
    properties:
      key:
        type: string
      value:
        type: string
    required:
    - key
    - value
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.ProductAttributeResponse:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.ProductResponse:
    properties:
      cartItems:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CartItemResponse'
        type: array
      category:
        $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CategoryResponse'
      description:
        type: string
      imageUrl:
        type: string
      isActive:
        type: boolean
      isFavorite:
        type: boolean
      name:
        type: string
      price:
        type: number
      productAttributes:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductAttributeResponse'
        type: array
      productStocks:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductStockResponse'
        type: array
      reviews:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ReviewResponse'
        type: array
      satisfactionRate:
        type: number
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.ProductStockResponse:
    properties:
      quantity:
        type: integer
      sellerId:
        type: integer
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.ReviewResponse:
    properties:
      content:
        type: string
      customerId:
        type: integer
      rating:
        type: integer
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
  github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_validation.ValidationError'
        type: array
    type: object
  github_com_wahid-wex_e-commerce-backend_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 42901
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - LimiterError
    - CustomRecovery
    - InternalError
  github_com_wahid-wex_e-commerce-backend_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/products/:
    post:
      consumes:
      - application/json
      description: Create a Product
      parameters:
      - description: Create a Product
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateProductRequest'
      responses:
        "201":
          description: Product response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a Product
      tags:
      - Products
  /v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Product
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a Product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get a Product
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Product response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a Product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update a Product
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Product
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateProductRequest'
      responses:
        "200":
          description: Product response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.ProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a Product
      tags:
      - Products
  /v1/products/change-stock:
    post:
      consumes:
      - application/json
      description: Change or create stock of product
      parameters:
      - description: Stock Change
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CreateUpdateProductStockRequest'
      responses:
        "200":
          description: Stock response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Change Stock
      tags:
      - Products
  /v1/products/favorite:
    delete:
      consumes:
      - application/json
      description: remove a product from customer's favorites
      parameters:
      - description: Add to favorites request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.AddRemoveToFavoriteRequest'
      responses:
        "200":
          description: Favorite response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.FavoriteResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: remove a product from favorites
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Add a product to customer's favorites
      parameters:
      - description: Add to favorites request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.AddRemoveToFavoriteRequest'
      responses:
        "200":
          description: Favorite response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.FavoriteResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Add a product to favorites
      tags:
      - Products
  /v1/products/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get a Product
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: Product response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.PagedList-github_com_wahid-wex_e-commerce-backend_api_dto_ProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a Product
      tags:
      - Products
  /v1/products/leave-review:
    post:
      consumes:
      - application/json
      description: Add a product to customer's favorites
      parameters:
      - description: Add to favorites request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.CreateReviewRequest'
      responses:
        "200":
          description: Favorite response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
            - properties:
                result:
                  type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Add a product to favorites
      tags:
      - Products
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"